refactor: relocate Reservation types to services layer and purify dependency graph (X-Tracker #350)
Move Reservation and ReservationMessage types out of components directory and centralize them inside src/services/reservations/types.ts to enforce unidirectional dependency guidelines.

Ref: https://github.com/Xchange-Taiwan/X-Talent-Tracker/issues/350
