+++ src/hooks/user/reservation/useReservationData.ts
@@ -1,10 +1,10 @@
 import { useSession } from 'next-auth/react';
 import { useCallback, useEffect, useRef, useState } from 'react';
 
-import { Reservation } from '@/components/reservation/types';
 import { trackEvent } from '@/lib/analytics';
 import { captureFlowFailure } from '@/lib/monitoring';
 import { fetchReservations, ReservationState } from '@/services/reservations';
+import { Reservation } from '@/services/reservations/types';
 
 export type ReservationRole = 'mentee' | 'mentor';
